Simplify deallocation
We should still free it at the end so that tools that test for
memory leaks won't have a positive that could obscure others, and
because it clearly expresses that there *was* an allocation, etc.

But zeroing the pointer variable out here is not necessary. That
macro was introduced for a different situation where freeing wasn't
the last action before the end of the block where variables were
defined. The code that really benefited from it was previously
removed.